Refactor deep link recipe packages
Group the existing recipes on how to handle deep links into their own "handlerequests" package. This differentiates them from the "usecases" package in pr#313 for recipes on custom deep link components.
