+++ app/src/main/java/com/example/nav3recipes/deeplink/handlerequests/staticuri/README.md
